Refine switch warnings
Adjust compiler options:

- don't warn if not all enum values are covered, but there is default
  branch
- warn if not all enum values are covered, and there is no default
  branch

Remove now useless (and harmless) casts in switches.

+++ src/internal_modules/roc_address/addr_family.cpp
@@ -12,7 +12,7 @@ namespace roc {
 namespace address {
 
 const char* addr_family_to_str(AddrFamily family) {
-    switch ((int)family) {
+    switch (family) {
     case Family_IPv4:
         return "IPv4";
